+++ clients/all.d.ts
@@ -344,3 +344,5 @@ export import IVSRealTime = require('./ivsrealtime');
 export import VPCLattice = require('./vpclattice');
 export import OSIS = require('./osis');
 export import MediaPackageV2 = require('./mediapackagev2');
+export import PaymentCryptography = require('./paymentcryptography');
+export import PaymentCryptographyData = require('./paymentcryptographydata');